src/back/space/tests/requirements.txt
yandex-taxi-testsuite >= 0.1.17
websockets >= 11.0.3
psycopg2 == 2.9.9
